reimplement lightning rod
-lightning rods now emit particles during a thunderstorm if they can attract lightning
-if a lightning rod attracts lightning, it will no longer clear existing fire around where the lightning would have landed
-ice stone is no longer rendered as a multi-layer block

+++ src/main/java/git/jbredwards/njarm/mod/common/util/AABBUtils.java
@@ -13,8 +13,8 @@ import javax.annotation.Nonnull;
 public final class AABBUtils
 {
     @Nonnull
-    public static AxisAlignedBB of(int minX, int minY, int minZ, int maxX, int maxY, int maxZ) {
-        return new AxisAlignedBB(minX / 16.0, minY / 16.0, minZ / 16.0, maxX / 16.0, maxY / 16.0, maxZ / 16.0);
+    public static AxisAlignedBB of(double minX, double minY, double minZ, double maxX, double maxY, double maxZ) {
+        return new AxisAlignedBB(minX / 16, minY / 16, minZ / 16, maxX / 16, maxY / 16, maxZ / 16);
     }
 
     @Nonnull
